+++ browser.js
@@ -1,6 +1,12 @@
 import { FetchSession } from './fetch.js';
+import { mixCombiningSessionInto } from './combine.js';
 
 export default class BrowserSession extends FetchSession {
 }
 
-export { ApiErrors } from './core.js';
+mixCombiningSessionInto( BrowserSession );
+
+export {
+	ApiErrors,
+	set,
+} from './core.js';