+++ node.js
@@ -1,6 +1,12 @@
 import { AxiosSession } from './axios.js';
+import { mixCombiningSessionInto } from './combine.js';
 
 export default class NodeSession extends AxiosSession {
 }
 
-export { ApiErrors } from './core.js';
+mixCombiningSessionInto( NodeSession );
+
+export {
+	ApiErrors,
+	set,
+} from './core.js';